requirements.txt
numpy == 2.3.5
matplotlib == 3.10.8
pandas == 2.3.3
pysam == 0.23.3
scipy == 1.16.3
python-dotenv == 1.2.1
biopython == 1.86
bcbio-gff == 0.7.1